More NPC work
TODO:
- Fix issue with searchable entities and cake npc orb giving. As it is
now there is an “infinite item” glitch that allows the player to get
multiple of the same item. For example, searching a tree should remove
the item so that the player cannot pick it up again. Also the npc that
needs the cake delivered can give 2 orbs instead of just one. This
needs to be fixed.